src/lib/p2p/mod.rs:add Network
